Add --no-cache option to wget commands
Updated wget commands to include --no-cache option for better file retrieval.

+++ fnancz-installer2.sh
@@ -1,17 +1,30 @@
+scriptlocation=$(pwd -L)
+scriptdeletion() {
+rm -f "$scriptlocation/fnancz2-installer1.sh"
+}
+trap scriptdeletion EXIT
+
 echo -e "\e[31mP.S: you must give full path of the zip not a shortcut\e[0m"
 read -r -p "Where is the path that you have the zip of the game downloaded? Defaulted=Downloads:" path
+gamename="Five Nights at NCZ AE"
+
 cd "$HOME/Downloads"
 cd "$path" # kinda stupid way that handles the operation but eh works for now
 gamezip=$(ls -v -r 'Five Nights at NCZ AE v'* | head -n 1)
 gamezip2=$(ls -v -r 'Five+Nights+at+NCZ+AE+v'* | head -n 1)
-gamename="Five Nights at NCZ AE"
+mv "$gamezip2" "$gamezip" >/dev/null 2>&1
+if [ ! -f "$gamezip" ]; then
+    echo -e "\e[31mError: the game zip file doesn't exist in $(pwd -L)!\e[0m"
+    exit 1
+
+fi
+
 
 if [ -d "$gamename" ]; then
     echo -e "\e[31mError: '$gamename' folder already exists at this location!\e[0m"
     exit 1
 fi
 
-mv "$gamezip2" "$gamezip" >/dev/null 2>&1
 mkdir "$gamename"
 cp "$gamezip" "$gamename"
 cd "$gamename"
@@ -24,11 +37,11 @@ mv "6am_animation.mp4" "assets"
 mv "options.ini" "assets"
 mv "assets/data.win" "assets/game.unx"
 rm NekoPresence_x64.dll
-wget https://raw.githubusercontent.com/AmrThePigeon/FNANCZAE1_Script_Builder/refs/heads/main/Five_Nights_at_NCZ_AE
-wget https://raw.githubusercontent.com/AmrThePigeon/FNANCZAE1_Script_Builder/refs/heads/main/lib.zip
-wget https://raw.githubusercontent.com/AmrThePigeon/FNANCZAE1_Script_Builder/refs/heads/main/run.sh
-wget https://raw.githubusercontent.com/AmrThePigeon/FNANCZAE1_Script_Builder/refs/heads/main/libNekoPresence_x64_linux_arm64-v8a.so
-wget https://raw.githubusercontent.com/AmrThePigeon/FNANCZAE1_Script_Builder/refs/heads/main/icon.png
+wget https://raw.githubusercontent.com/AmrThePigeon/FNANCZAE1_Script_Builder/refs/heads/main/Five_Nights_at_NCZ_AE --no-cache
+wget https://raw.githubusercontent.com/AmrThePigeon/FNANCZAE1_Script_Builder/refs/heads/main/lib.zip --no-cache
+wget https://raw.githubusercontent.com/AmrThePigeon/FNANCZAE1_Script_Builder/refs/heads/main/run.sh --no-cache
+wget https://raw.githubusercontent.com/AmrThePigeon/FNANCZAE1_Script_Builder/refs/heads/main/libNekoPresence_x64_linux_arm64-v8a.so --no-cache
+wget https://raw.githubusercontent.com/AmrThePigeon/FNANCZAE1_Script_Builder/refs/heads/main/icon.png --no-cache
 mv icon.png assets
 mv libNekoPresence_x64_linux_arm64-v8a.so assets
 unzip lib.zip
